+++ lib/rust/profiler/src/lib.rs
@@ -155,6 +155,7 @@ use std::str;
 use internal::*;
 
 
+
 // ======================
 // === MetadataLogger ===
 // ======================
